remove redundant widgets
